moved telegram fulfillment to telegram controller
